web/requirements.txt
django == 3.1.2
open3d == 0.10.0
scikit-learn == 0.23.2
descartes == 1.1.0
numpy == 1.19.2
alphashape == 1.0.2
pywavefront == 1.3.2
pandas
djangorestframework==3.11.0
